+++ test/unit/fuzz-tools.test.js
@@ -240,6 +240,7 @@ import * as pedstoxV247 from '../../lib/pedstox-v247.js';
 import * as woundidV248 from '../../lib/woundid-v248.js';
 import * as renalpulmV249 from '../../lib/renalpulm-v249.js';
 import * as obgynV250 from '../../lib/obgyn-v250.js';
+import * as cardiometabV251 from '../../lib/cardiometab-v251.js';
 
 const MODULES = {
   'clinical.js': clinical,
@@ -424,6 +425,7 @@ const MODULES = {
   'woundid-v248.js': woundidV248,
   'renalpulm-v249.js': renalpulmV249,
   'obgyn-v250.js': obgynV250,
+  'cardiometab-v251.js': cardiometabV251,
 };
 
 const MATRIX = [0, -1, 1e9, NaN, Infinity, -Infinity, '', undefined, null];